refactor(channels): move discord_delivery_provider -> channels/discord/delivery_provider
Slice 4 of the phase-1a lane. Provider adapter is channel mechanics per the
boundary doctrine (provider-specific exception mapping stays at the adapter).
One lazy-import line in #3239-contested discord-bridge.py, resolved at the
post-merge rebase.
